# docs/reference/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "keepkey-sdk-server",
    "version": "0.1.1",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://localhost:1646/"
    }
  ],
  "paths": {
...
    "/addresses/eth": {
      "post": {
        "operationId": "EthereumGetAddress",
        "summary": "Get Ethereum Address",
        "description": "Get an Ethereum address, loading it into the current session and optionally displaying it on-device.",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "show_display": {
                    "type": "boolean"
                  },
                  "address_n": {
                    "$ref": "#/components/schemas/types.AddressNList"
                  }
                },
                "required": [
                  "address_n"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "address": {
                      "$ref": "#/components/schemas/types.eth.Address"
                    }
                  },
                  "required": [
                    "address"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Error processing request"
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "tags": [
          "Address"
        ]
      }
    },
...
  },
  "components": {
...
    "schemas": {
      "types.numeric.UnsignedInteger.Safe": {
        "type": "integer",
        "format": "int64",
        "maximum": 9007199254740991,
        "minimum": 0
      },
      "types.numeric.U32": {
        "format": "uint32",
        "allOf": [
          {
            "$ref": "#/components/schemas/types.numeric.UnsignedInteger.Safe"
          },
          {}
        ],
        "maximum": 4294967295,
        "minimum": 0
      },
      "types.AddressNList": {
        "description": "a BIP-32 path, expressed as a list of node indexes",
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/types.numeric.U32"
        },
        "maxItems": 5,
        "minItems": 3
      },
...
      "types.eth.Address": {
        "description": "an ethereum address, possibly checksummed",
        "type": "string",
        "format": "eth_data",
        "maxLength": 42,
        "minLength": 42,
        "pattern": "^0x(?:[0-9a-fA-F]{2}){20}$"
      },
...
    },
...
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "SESSIONID"
      }
    }
  }
}